1) Put common /mpitype/ back in dmc/dmc_elec/MPI/maindmc_mov1_mpi.f for mpi3 to work. 2) Correct the bounds so that mpi2 and mpi3 work.
Note: mpi2 and mpi3 ought to give identical energies since they differ on in one routine,
the difference being that mpi3 does not send unnecessary jastrow data between processors.
In a very old version of the program, before F90 changes were made, mpi2 and mpi3 do give
identical results as expected.  However, in this version there seems to be a build up of
round-off error, but I really see no reason for this to occur.



git-svn-id: svn://hydrogene.lct.jussieu.fr/champ@315 36b7402c-5635-0410-8f0e-a47a0fc0da44